+++ flake.nix
@@ -4,15 +4,26 @@
     rust-overlay.url = "github:oxalica/rust-overlay";
     rust-overlay.inputs.nixpkgs.follows = "nixpkgs";
     flake-utils.url = "github:numtide/flake-utils";
+    untangle.url = "github:jonochang/untangle";
+    untangle.inputs.nixpkgs.follows = "nixpkgs";
   };
 
-  outputs = { self, nixpkgs, rust-overlay, flake-utils }:
+  outputs = { self, nixpkgs, rust-overlay, flake-utils, untangle }:
     flake-utils.lib.eachDefaultSystem (system:
       let
         pkgs = import nixpkgs {
           inherit system;
           overlays = [ rust-overlay.overlays.default ];
         };
+        untanglePkg = untangle.packages.${system}.default.overrideAttrs (old: {
+          src = pkgs.fetchFromGitHub {
+            owner = "jonochang";
+            repo = "untangle";
+            rev = "v${old.version}";
+            hash = "sha256-zHhk6f50bjiRi0PnY3YWcVvlzhj8q1NLroPxDVYOP7o=";
+          };
+          doCheck = false;
+        });
         rustToolchain = pkgs.rust-bin.stable.latest.default.override {
           extensions = [ "clippy" "rustfmt" "rust-src" ];
         };
@@ -30,6 +41,7 @@
             pkgs.cargo-nextest pkgs.cargo-deny
             pkgs.cargo-llvm-cov pkgs.cargo-insta
             pkgs.mdbook pkgs.git
+            untanglePkg
           ];
           LIBGIT2_NO_VENDOR = "1";
           OPENSSL_DIR = "${pkgs.openssl.dev}";